pyproject.toml
[project]
name = "baml-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "baml-py>=0.87.2",
    "dotenv>=0.9.9",
    "google-adk>=0.5.0",
    "pydantic>=2.11.4",
]
